+++ src/managednetworkfabric/azext_managednetworkfabric/aaz/latest/networkfabric/bootstrapdevice/__init__.py
@@ -10,7 +10,11 @@
 
 from .__cmd_group import *
 from ._list import *
+from ._reboot import *
+from ._refresh_config import *
+from ._resync_password import *
 from ._run_ro_command import *
 from ._run_rw_command import *
 from ._show import *
+from ._update_admin_state import *
 from ._wait import *