+++ src/managednetworkfabric/azext_managednetworkfabric/aaz/latest/networkfabric/bootstrapdevice/_resync_password.py
@@ -12,19 +12,19 @@ from azure.cli.core.aaz import *
 
 
 @register_command(
-    "networkfabric fabric commit-configuration",
+    "networkfabric bootstrapdevice resync-password",
 )
-class CommitConfiguration(AAZCommand):
-    """Atomic update of the given Network Fabric instance. Sync update of NFA resources at Fabric level.
+class ResyncPassword(AAZCommand):
+    """Updates the Network Bootstrap Device to use the latest passwords. Does not generate new passwords. Allows network bootstrap devices missed during a previous password rotation to be brought back into sync.
 
-    :example: Run commit configuration on the Network Fabric
-        az networkfabric fabric commit-configuration --resource-group "example-rg" --resource-name "example-fabric"
+    :example: Resync the latest passwords to the Network Bootstrap Device
+        az networkfabric bootstrapdevice resync-password --resource-group example-rg --resource-name example-device
     """
 
     _aaz_info = {
         "version": "2026-07-15-preview",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.managednetworkfabric/networkfabrics/{}/commitconfiguration", "2026-07-15-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.managednetworkfabric/networkbootstrapdevices/{}/resyncpasswords", "2026-07-15-preview"],
         ]
     }
 
@@ -47,7 +47,7 @@ class CommitConfiguration(AAZCommand):
         _args_schema = cls._args_schema
         _args_schema.resource_name = AAZStrArg(
             options=["--resource-name"],
-            help="Name of the Network Fabric.",
+            help="Name of the Network Bootstrap Device.",
             required=True,
             id_part="name",
             fmt=AAZStrArgFormat(
@@ -57,35 +57,11 @@ class CommitConfiguration(AAZCommand):
         _args_schema.resource_group = AAZResourceGroupNameArg(
             required=True,
         )
-
-        # define Arg Group "Body"
-
-        _args_schema = cls._args_schema
-        _args_schema.commit_policy = AAZStrArg(
-            options=["--commit-policy"],
-            arg_group="Body",
-            help="Commit configuration Policy. Supported policy is StageCEConfiguration, which indicates to prepare the configuration for the CE device type.",
-            enum={"StageCEConfiguration": "StageCEConfiguration"},
-        )
-        _args_schema.commit_stage = AAZStrArg(
-            options=["--commit-stage"],
-            arg_group="Body",
-            help="Commit stage Action to be performed.",
-            enum={"Continue": "Continue", "Rollback": "Rollback", "Start": "Start"},
-        )
-        _args_schema.devices = AAZListArg(
-            options=["--devices"],
-            arg_group="Body",
-            help="List of ARM resource IDs of devices to be included in the commit operation. Either CE1 or CE2 is allowed.",
-        )
-
-        devices = cls._args_schema.devices
-        devices.Element = AAZStrArg()
         return cls._args_schema
 
     def _execute_operations(self):
         self.pre_operations()
-        yield self.NetworkFabricsCommitConfiguration(ctx=self.ctx)()
+        yield self.NetworkBootstrapDevicesResyncPasswords(ctx=self.ctx)()
         self.post_operations()
 
     @register_callback
@@ -100,7 +76,7 @@ class CommitConfiguration(AAZCommand):
         result = self.deserialize_output(self.ctx.vars.instance, client_flatten=True)
         return result
 
-    class NetworkFabricsCommitConfiguration(AAZHttpOperation):
+    class NetworkBootstrapDevicesResyncPasswords(AAZHttpOperation):
         CLIENT_TYPE = "MgmtClient"
 
         def __call__(self, *args, **kwargs):
@@ -130,7 +106,7 @@ class CommitConfiguration(AAZCommand):
         @property
         def url(self):
             return self.client.format_url(
-                "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.ManagedNetworkFabric/networkFabrics/{networkFabricName}/commitConfiguration",
+                "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.ManagedNetworkFabric/networkBootstrapDevices/{networkBootstrapDeviceName}/resyncPasswords",
                 **self.url_parameters
             )
 
@@ -146,7 +122,7 @@ class CommitConfiguration(AAZCommand):
         def url_parameters(self):
             parameters = {
                 **self.serialize_url_param(
-                    "networkFabricName", self.ctx.args.resource_name,
+                    "networkBootstrapDeviceName", self.ctx.args.resource_name,
                     required=True,
                 ),
                 **self.serialize_url_param(
@@ -174,30 +150,10 @@ class CommitConfiguration(AAZCommand):
         def header_parameters(self):
             parameters = {
                 **self.serialize_header_param(
-                    "Content-Type", "application/json",
-                ),
-                **self.serialize_header_param(
                     "Accept", "application/json",
                 ),
             }
             return parameters
-
-        @property
-        def content(self):
-            _content_value, _builder = self.new_content_builder(
-                self.ctx.args,
-                typ=AAZObjectType,
-                typ_kwargs={"flags": {"client_flatten": True}}
-            )
-            _builder.set_prop("commitPolicy", AAZStrType, ".commit_policy")
-            _builder.set_prop("commitStage", AAZStrType, ".commit_stage")
-            _builder.set_prop("devices", AAZListType, ".devices")
-
-            devices = _builder.get(".devices")
-            if devices is not None:
-                devices.set_elements(AAZStrType, ".")
-
-            return self.serialize_content(_content_value)
 
         def on_200(self, session):
             data = self.deserialize_http_content(session)
@@ -221,7 +177,7 @@ class CommitConfiguration(AAZCommand):
                 serialized_name="endTime",
             )
             _schema_on_200.error = AAZObjectType()
-            _CommitConfigurationHelper._build_schema_error_detail_read(_schema_on_200.error)
+            _ResyncPasswordHelper._build_schema_error_detail_read(_schema_on_200.error)
             _schema_on_200.id = AAZStrType(
                 nullable=True,
             )
@@ -244,13 +200,13 @@ class CommitConfiguration(AAZCommand):
 
             operations = cls._schema_on_200.operations
             operations.Element = AAZObjectType()
-            _CommitConfigurationHelper._build_schema_operation_status_result_read(operations.Element)
+            _ResyncPasswordHelper._build_schema_operation_status_result_read(operations.Element)
 
             return cls._schema_on_200
 
 
-class _CommitConfigurationHelper:
-    """Helper class for CommitConfiguration"""
+class _ResyncPasswordHelper:
+    """Helper class for ResyncPassword"""
 
     _schema_error_detail_read = None
 
@@ -364,4 +320,4 @@ class _CommitConfigurationHelper:
         _schema.status = cls._schema_operation_status_result_read.status
 
 
-__all__ = ["CommitConfiguration"]
+__all__ = ["ResyncPassword"]