+++ src/managednetworkfabric/azext_managednetworkfabric/aaz/latest/networkfabric/bootstrapinterface/__init__.py
@@ -11,4 +11,5 @@
 from .__cmd_group import *
 from ._list import *
 from ._show import *
+from ._update_admin_state import *
 from ._wait import *