+++ src/managednetworkfabric/azext_managednetworkfabric/aaz/latest/networkfabric/fabric/_commit_config.py
@@ -12,19 +12,19 @@ from azure.cli.core.aaz import *
 
 
 @register_command(
-    "networkfabric device reboot",
+    "networkfabric fabric commit-config",
 )
-class Reboot(AAZCommand):
-    """Reboot the Network Device.
+class CommitConfig(AAZCommand):
+    """Atomic update of the given Network Fabric instance. Sync update of NFA resources at Fabric level.
 
-    :example: Reboot the Network Device
-        az networkfabric device reboot --resource-group example-rg --resource-name example-device --reboot-type GracefulRebootWithZTP
+    :example: Run commit configuration on the Network Fabric
+        az networkfabric fabric commit-config--resource-group "example-rg" --resource-name "example-fabric"
     """
 
     _aaz_info = {
         "version": "2026-07-15-preview",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.managednetworkfabric/networkdevices/{}/reboot", "2026-07-15-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.managednetworkfabric/networkfabrics/{}/commitconfiguration", "2026-07-15-preview"],
         ]
     }
 
@@ -47,7 +47,7 @@ class Reboot(AAZCommand):
         _args_schema = cls._args_schema
         _args_schema.resource_name = AAZStrArg(
             options=["--resource-name"],
-            help="Name of the Network Device.",
+            help="Name of the Network Fabric.",
             required=True,
             id_part="name",
             fmt=AAZStrArgFormat(
@@ -61,17 +61,31 @@ class Reboot(AAZCommand):
         # define Arg Group "Body"
 
         _args_schema = cls._args_schema
-        _args_schema.reboot_type = AAZStrArg(
-            options=["--reboot-type"],
+        _args_schema.commit_policy = AAZStrArg(
+            options=["--commit-policy"],
             arg_group="Body",
-            help="Type of reboot to be performed. Example: GracefulRebootWithZTP",
-            enum={"GracefulRebootWithZTP": "GracefulRebootWithZTP", "GracefulRebootWithoutZTP": "GracefulRebootWithoutZTP", "UngracefulRebootWithZTP": "UngracefulRebootWithZTP", "UngracefulRebootWithoutZTP": "UngracefulRebootWithoutZTP"},
+            help="Commit configuration Policy. Supported policy is StageCEConfiguration, which indicates to prepare the configuration for the CE device type.",
+            enum={"StageCEConfiguration": "StageCEConfiguration"},
         )
+        _args_schema.commit_stage = AAZStrArg(
+            options=["--commit-stage"],
+            arg_group="Body",
+            help="Commit stage Action to be performed.",
+            enum={"Continue": "Continue", "Rollback": "Rollback", "Start": "Start"},
+        )
+        _args_schema.devices = AAZListArg(
+            options=["--devices"],
+            arg_group="Body",
+            help="List of ARM resource IDs of devices to be included in the commit operation. Either CE1 or CE2 is allowed.",
+        )
+
+        devices = cls._args_schema.devices
+        devices.Element = AAZStrArg()
         return cls._args_schema
 
     def _execute_operations(self):
         self.pre_operations()
-        yield self.NetworkDevicesReboot(ctx=self.ctx)()
+        yield self.NetworkFabricsCommitConfiguration(ctx=self.ctx)()
         self.post_operations()
 
     @register_callback
@@ -86,7 +100,7 @@ class Reboot(AAZCommand):
         result = self.deserialize_output(self.ctx.vars.instance, client_flatten=True)
         return result
 
-    class NetworkDevicesReboot(AAZHttpOperation):
+    class NetworkFabricsCommitConfiguration(AAZHttpOperation):
         CLIENT_TYPE = "MgmtClient"
 
         def __call__(self, *args, **kwargs):
@@ -116,7 +130,7 @@ class Reboot(AAZCommand):
         @property
         def url(self):
             return self.client.format_url(
-                "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.ManagedNetworkFabric/networkDevices/{networkDeviceName}/reboot",
+                "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.ManagedNetworkFabric/networkFabrics/{networkFabricName}/commitConfiguration",
                 **self.url_parameters
             )
 
@@ -132,7 +146,7 @@ class Reboot(AAZCommand):
         def url_parameters(self):
             parameters = {
                 **self.serialize_url_param(
-                    "networkDeviceName", self.ctx.args.resource_name,
+                    "networkFabricName", self.ctx.args.resource_name,
                     required=True,
                 ),
                 **self.serialize_url_param(
@@ -173,9 +187,15 @@ class Reboot(AAZCommand):
             _content_value, _builder = self.new_content_builder(
                 self.ctx.args,
                 typ=AAZObjectType,
-                typ_kwargs={"flags": {"required": True, "client_flatten": True}}
+                typ_kwargs={"flags": {"client_flatten": True}}
             )
-            _builder.set_prop("rebootType", AAZStrType, ".reboot_type")
+            _builder.set_prop("commitPolicy", AAZStrType, ".commit_policy")
+            _builder.set_prop("commitStage", AAZStrType, ".commit_stage")
+            _builder.set_prop("devices", AAZListType, ".devices")
+
+            devices = _builder.get(".devices")
+            if devices is not None:
+                devices.set_elements(AAZStrType, ".")
 
             return self.serialize_content(_content_value)
 
@@ -195,13 +215,42 @@ class Reboot(AAZCommand):
                 return cls._schema_on_200
 
             cls._schema_on_200 = AAZObjectType()
-            _RebootHelper._build_schema_operation_status_result_read(cls._schema_on_200)
+
+            _schema_on_200 = cls._schema_on_200
+            _schema_on_200.end_time = AAZStrType(
+                serialized_name="endTime",
+            )
+            _schema_on_200.error = AAZObjectType()
+            _CommitConfigHelper._build_schema_error_detail_read(_schema_on_200.error)
+            _schema_on_200.id = AAZStrType(
+                nullable=True,
+            )
+            _schema_on_200.name = AAZStrType()
+            _schema_on_200.operations = AAZListType()
+            _schema_on_200.percent_complete = AAZFloatType(
+                serialized_name="percentComplete",
+            )
+            _schema_on_200.resource_id = AAZStrType(
+                serialized_name="resourceId",
+                nullable=True,
+                flags={"read_only": True},
+            )
+            _schema_on_200.start_time = AAZStrType(
+                serialized_name="startTime",
+            )
+            _schema_on_200.status = AAZStrType(
+                flags={"required": True},
+            )
+
+            operations = cls._schema_on_200.operations
+            operations.Element = AAZObjectType()
+            _CommitConfigHelper._build_schema_operation_status_result_read(operations.Element)
 
             return cls._schema_on_200
 
 
-class _RebootHelper:
-    """Helper class for Reboot"""
+class _CommitConfigHelper:
+    """Helper class for CommitConfig"""
 
     _schema_error_detail_read = None
 
@@ -315,4 +364,4 @@ class _RebootHelper:
         _schema.status = cls._schema_operation_status_result_read.status
 
 
-__all__ = ["Reboot"]
+__all__ = ["CommitConfig"]