+++ src/managednetworkfabric/azext_managednetworkfabric/aaz/latest/networkfabric/fabric/_view_device_config.py
@@ -12,19 +12,19 @@ from azure.cli.core.aaz import *
 
 
 @register_command(
-    "networkfabric device refresh-configuration",
+    "networkfabric fabric view-device-config",
 )
-class RefreshConfiguration(AAZCommand):
-    """Refreshes the configuration the Network Device.
+class ViewDeviceConfig(AAZCommand):
+    """Post action: Triggers view of network fabric configuration.
 
-    :example: Run refresh configuration on the Network Device
-        az networkfabric device refresh-configuration --resource-group example-rg --resource-name example-device
+    :example: View device configuration on the Network Fabric
+        az networkfabric fabric view-device-config --resource-group example-rg --resource-name example-fabric
     """
 
     _aaz_info = {
         "version": "2026-07-15-preview",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.managednetworkfabric/networkdevices/{}/refreshconfiguration", "2026-07-15-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.managednetworkfabric/networkfabrics/{}/viewdeviceconfiguration", "2026-07-15-preview"],
         ]
     }
 
@@ -45,9 +45,9 @@ class RefreshConfiguration(AAZCommand):
         # define Arg Group ""
 
         _args_schema = cls._args_schema
-        _args_schema.network_device_name = AAZStrArg(
-            options=["--resource-name", "--network-device-name"],
-            help="Name of the Network Device.",
+        _args_schema.network_fabric_name = AAZStrArg(
+            options=["--resource-name", "--network-fabric-name"],
+            help="Name of the Network Fabric.",
             required=True,
             id_part="name",
             fmt=AAZStrArgFormat(
@@ -61,7 +61,7 @@ class RefreshConfiguration(AAZCommand):
 
     def _execute_operations(self):
         self.pre_operations()
-        yield self.NetworkDevicesRefreshConfiguration(ctx=self.ctx)()
+        yield self.NetworkFabricsViewDeviceConfiguration(ctx=self.ctx)()
         self.post_operations()
 
     @register_callback
@@ -76,7 +76,7 @@ class RefreshConfiguration(AAZCommand):
         result = self.deserialize_output(self.ctx.vars.instance, client_flatten=True)
         return result
 
-    class NetworkDevicesRefreshConfiguration(AAZHttpOperation):
+    class NetworkFabricsViewDeviceConfiguration(AAZHttpOperation):
         CLIENT_TYPE = "MgmtClient"
 
         def __call__(self, *args, **kwargs):
@@ -106,7 +106,7 @@ class RefreshConfiguration(AAZCommand):
         @property
         def url(self):
             return self.client.format_url(
-                "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.ManagedNetworkFabric/networkDevices/{networkDeviceName}/refreshConfiguration",
+                "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.ManagedNetworkFabric/networkFabrics/{networkFabricName}/viewDeviceConfiguration",
                 **self.url_parameters
             )
 
@@ -122,7 +122,7 @@ class RefreshConfiguration(AAZCommand):
         def url_parameters(self):
             parameters = {
                 **self.serialize_url_param(
-                    "networkDeviceName", self.ctx.args.network_device_name,
+                    "networkFabricName", self.ctx.args.network_fabric_name,
                     required=True,
                 ),
                 **self.serialize_url_param(
@@ -177,7 +177,7 @@ class RefreshConfiguration(AAZCommand):
                 serialized_name="endTime",
             )
             _schema_on_200.error = AAZObjectType()
-            _RefreshConfigurationHelper._build_schema_error_detail_read(_schema_on_200.error)
+            _ViewDeviceConfigHelper._build_schema_error_detail_read(_schema_on_200.error)
             _schema_on_200.id = AAZStrType(
                 nullable=True,
             )
@@ -186,6 +186,7 @@ class RefreshConfiguration(AAZCommand):
             _schema_on_200.percent_complete = AAZFloatType(
                 serialized_name="percentComplete",
             )
+            _schema_on_200.properties = AAZObjectType()
             _schema_on_200.resource_id = AAZStrType(
                 serialized_name="resourceId",
                 nullable=True,
@@ -200,13 +201,18 @@ class RefreshConfiguration(AAZCommand):
 
             operations = cls._schema_on_200.operations
             operations.Element = AAZObjectType()
-            _RefreshConfigurationHelper._build_schema_operation_status_result_read(operations.Element)
+            _ViewDeviceConfigHelper._build_schema_operation_status_result_read(operations.Element)
+
+            properties = cls._schema_on_200.properties
+            properties.device_configuration_url = AAZStrType(
+                serialized_name="deviceConfigurationUrl",
+            )
 
             return cls._schema_on_200
 
 
-class _RefreshConfigurationHelper:
-    """Helper class for RefreshConfiguration"""
+class _ViewDeviceConfigHelper:
+    """Helper class for ViewDeviceConfig"""
 
     _schema_error_detail_read = None
 
@@ -320,4 +326,4 @@ class _RefreshConfigurationHelper:
         _schema.status = cls._schema_operation_status_result_read.status
 
 
-__all__ = ["RefreshConfiguration"]
+__all__ = ["ViewDeviceConfig"]